+++ cli/src/pcluster/models/cluster.py
@@ -12,6 +12,7 @@
 # This module contains all the classes representing the Resources objects.
 # These objects are obtained from the configuration file through a conversion based on the Schema classes.
 #
+import hashlib
 import json
 import logging
 import os
@@ -21,14 +22,16 @@ from copy import deepcopy
 from datetime import datetime
 from enum import Enum
 from typing import List, Optional, Set, Tuple
+from urllib.request import urlopen
 
 import pkg_resources
 import yaml
+from jinja2 import BaseLoader, Environment
 from marshmallow import ValidationError
 
 from pcluster.aws.aws_api import AWSApi
 from pcluster.aws.common import AWSClientError, BadRequestError, LimitExceededError, StackNotFoundError, get_region
-from pcluster.config.cluster_config import BaseClusterConfig, SlurmScheduling, Tag
+from pcluster.config.cluster_config import BaseClusterConfig, ByosScheduling, SlurmScheduling, Tag
 from pcluster.config.common import ValidatorSuppressor
 from pcluster.config.config_patch import ConfigPatch
 from pcluster.constants import (
@@ -59,10 +62,10 @@ from pcluster.models.common import (
     upload_archive,
 )
 from pcluster.models.compute_fleet_status_manager import ComputeFleetStatus, ComputeFleetStatusManager
-from pcluster.models.s3_bucket import S3Bucket, S3BucketFactory, S3FileFormat, create_s3_presigned_url
+from pcluster.models.s3_bucket import S3Bucket, S3BucketFactory, S3FileFormat, create_s3_presigned_url, parse_bucket_url
 from pcluster.schemas.cluster_schema import ClusterSchema
 from pcluster.templates.cdk_builder import CDKTemplateBuilder
-from pcluster.utils import datetime_to_epoch, generate_random_name_with_prefix, get_installed_version, grouper
+from pcluster.utils import datetime_to_epoch, generate_random_name_with_prefix, get_attr, get_installed_version, grouper
 from pcluster.validators.common import FailureLevel, ValidationResult
 
 # pylint: disable=C0302
@@ -271,14 +274,14 @@ class Cluster:
     def compute_fleet_status_with_last_updated_time(self):
         """Status of the cluster compute fleet and the last compute fleet status updated time."""
         if self.stack.is_working_status or self.stack.status == "UPDATE_IN_PROGRESS":
-            if self.stack.scheduler == "slurm":
-                compute_fleet_status_manager = ComputeFleetStatusManager(self.name)
-                status, last_updated_time = compute_fleet_status_manager.get_status_with_last_updated_time()
-            else:  # scheduler is AWS Batch:
+            if self.stack.scheduler == "awsbatch":
                 status = ComputeFleetStatus(
                     AWSApi.instance().batch.get_compute_environment_state(self.stack.batch_compute_environment)
                 )
                 last_updated_time = None
+            else:
+                compute_fleet_status_manager = ComputeFleetStatusManager(self.name)
+                status, last_updated_time = compute_fleet_status_manager.get_status_with_last_updated_time()
             return status, last_updated_time
         else:
             LOGGER.info(
@@ -377,7 +380,7 @@ class Cluster:
                 self.bucket.delete_s3_artifacts()
             raise _cluster_error_mapper(e, str(e))
 
-    def _load_config(self, cluster_config: dict) -> ClusterSchema:
+    def _load_config(self, cluster_config: dict) -> BaseClusterConfig:
         """Load the config and catch / translate any errors that occur during loading."""
         try:
             return ClusterSchema(cluster_name=self.name).load(cluster_config)
@@ -483,17 +486,63 @@ class Cluster:
             if self.template_body:
                 self.bucket.upload_cfn_template(self.template_body, PCLUSTER_S3_ARTIFACTS_DICT.get("template_name"))
 
-            if isinstance(self.config.scheduling, SlurmScheduling):
+            if isinstance(self.config.scheduling, (SlurmScheduling, ByosScheduling)):
                 # upload instance types data
                 self.bucket.upload_config(
                     self.config.get_instance_types_data(),
                     PCLUSTER_S3_ARTIFACTS_DICT.get("instance_types_data_name"),
                     format=S3FileFormat.JSON,
                 )
+
+            if isinstance(self.config.scheduling, ByosScheduling):
+                self._render_and_upload_byos_template()
+        except BadRequestClusterActionError:
+            raise
         except Exception as e:
             message = f"Unable to upload cluster resources to the S3 bucket {self.bucket.name} due to exception: {e}"
             LOGGER.error(message)
             raise _cluster_error_mapper(e, message)
+
+    def _render_and_upload_byos_template(self):
+        byos_template = get_attr(
+            self.config, "scheduling.settings.scheduler_definition.cluster_infrastructure.cloud_formation.template"
+        )
+        if not byos_template:
+            return
+
+        try:
+            if byos_template.startswith("s3"):
+                bucket_parsing_result = parse_bucket_url(byos_template)
+                result = AWSApi.instance().s3.get_object(
+                    bucket_name=bucket_parsing_result["bucket_name"], key=bucket_parsing_result["object_key"]
+                )
+                file_content = result["Body"].read().decode("utf-8")
+            else:
+                with urlopen(byos_template) as f:  # nosec nosemgrep - byos_template url is properly validated
+                    file_content = f.read().decode("utf-8")
+        except Exception as e:
+            raise BadRequestClusterActionError(
+                f"Error while downloading scheduler plugin artifacts from '{byos_template}': {str(e)}"
+            ) from e
+
+        # jinja rendering
+        try:
+            environment = Environment(loader=BaseLoader)  # nosec nosemgrep
+            environment.filters["hash"] = (
+                lambda value: hashlib.sha1(value.encode()).hexdigest()[0:16].capitalize()  # nosec nosemgrep
+            )
+            template = environment.from_string(file_content)
+            rendered_template = template.render(
+                cluster_configuration=parse_config(self.source_config_text), cluster_name=self.name
+            )
+        except Exception as e:
+            raise BadRequestClusterActionError(
+                f"Error while rendering scheduler plugin template '{byos_template}': {str(e)}"
+            ) from e
+
+        self.bucket.upload_cfn_template(
+            rendered_template, PCLUSTER_S3_ARTIFACTS_DICT["byos_template_name"], S3FileFormat.TEXT
+        )
 
     def delete(self, keep_logs: bool = True):
         """Delete cluster preserving log groups."""