+++ cli/tests/pcluster/schemas/test_cluster_schema.py
@@ -16,8 +16,16 @@ import yaml
 from assertpy import assert_that
 from marshmallow.validate import ValidationError
 
-from pcluster.constants import MAX_NUMBER_OF_COMPUTE_RESOURCES, MAX_NUMBER_OF_QUEUES
+from pcluster.constants import SUPPORTED_OSES
 from pcluster.schemas.cluster_schema import (
+    ByosCloudFormationClusterInfrastructureSchema,
+    ByosClusterSharedArtifactSchema,
+    ByosFileSchema,
+    ByosLogsSchema,
+    ByosPluginResourcesSchema,
+    ByosSchedulerDefinitionSchema,
+    ByosSupportedDistrosSchema,
+    ByosUserSchema,
     ClusterSchema,
     HeadNodeIamSchema,
     ImageSchema,
@@ -25,6 +33,7 @@ from pcluster.schemas.cluster_schema import (
     SchedulingSchema,
     SharedStorageSchema,
 )
+from pcluster.utils import replace_url_parameters
 from tests.pcluster.aws.dummy_aws_api import mock_aws_api
 from tests.pcluster.utils import load_cluster_model_from_yaml
 
@@ -37,7 +46,9 @@ def _check_cluster_schema(config_file_name):
     cluster_schema = ClusterSchema(cluster_name="clustername")
     cluster_schema.context = {"delete_defaults_when_dump": True}
     output_json = cluster_schema.dump(cluster)
-    assert_that(json.dumps(input_yaml, sort_keys=True)).is_equal_to(json.dumps(output_json, sort_keys=True))
+    assert_that(replace_url_parameters(json.dumps(input_yaml, sort_keys=True))).is_equal_to(
+        json.dumps(output_json, sort_keys=True)
+    )
 
     # Print output yaml
     output_yaml = yaml.dump(output_json)
@@ -47,11 +58,21 @@ def _check_cluster_schema(config_file_name):
 @pytest.mark.parametrize("config_file_name", ["slurm.required.yaml", "slurm.full.yaml"])
 def test_cluster_schema_slurm(mocker, test_datadir, config_file_name):
     mock_aws_api(mocker)
+    mocker.patch("pcluster.utils.get_region", return_value="fake_region")
     _check_cluster_schema(config_file_name)
 
 
 @pytest.mark.parametrize("config_file_name", ["awsbatch.simple.yaml", "awsbatch.full.yaml"])
-def test_cluster_schema_awsbatch(test_datadir, config_file_name):
+def test_cluster_schema_awsbatch(mocker, test_datadir, config_file_name):
+    mock_aws_api(mocker)
+    mocker.patch("pcluster.utils.get_region", return_value="fake_region")
+    _check_cluster_schema(config_file_name)
+
+
+@pytest.mark.parametrize("config_file_name", ["byos.required.yaml", "byos.full.yaml"])
+def test_cluster_schema_byos(mocker, test_datadir, config_file_name):
+    mock_aws_api(mocker)
+    mocker.patch("pcluster.utils.get_region", return_value="fake_region")
     _check_cluster_schema(config_file_name)
 
 
@@ -239,26 +260,12 @@ def dummy_slurm_compute_resource(name, instance_type):
             },
             None,
         ),
-        (  # beyond maximum slurm queue length
-            {
-                "Scheduler": "slurm",
-                "SlurmQueues": dummpy_slurm_queue_list(11),
-            },
-            f"Queue.*Longer than maximum length {MAX_NUMBER_OF_QUEUES}",
-        ),
         (  # maximum slurm queue length
             {
                 "Scheduler": "slurm",
                 "SlurmQueues": [dummy_slurm_queue("queue1", number_of_compute_resource=5)],
             },
             None,
-        ),
-        (  # beyond maximum slurm queue length
-            {
-                "Scheduler": "slurm",
-                "SlurmQueues": [dummy_slurm_queue("queue1", number_of_compute_resource=6)],
-            },
-            f"ComputeResources.*Longer than maximum length {MAX_NUMBER_OF_COMPUTE_RESOURCES}",
         ),
     ],
 )
@@ -356,4 +363,207 @@ def test_scheduler_constraints_for_intel_packages(
         assert_that(cluster.scheduling.scheduler).is_equal_to(scheduler)
         assert_that(cluster.additional_packages.intel_software.intel_hpc_platform).is_equal_to(
             install_intel_packages_enabled
+        )
+
+
+@pytest.mark.parametrize(
+    "x86, arm64, failure_message",
+    [
+        (None, ["centos7"], None),
+        (["centos7"], None, None),
+        ("ubuntu1804", ["centos7"], "Not a valid list"),
+        (["ubuntu1804"], ["centos7"], None),
+    ],
+)
+def test_byos_supported_distros_schema(x86, arm64, failure_message):
+    byos_supported_distros_schema = {}
+    if x86:
+        byos_supported_distros_schema["X86"] = x86
+    if arm64:
+        byos_supported_distros_schema["Arm64"] = arm64
+
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosSupportedDistrosSchema().load(byos_supported_distros_schema)
+    else:
+        byos_supported_distros = ByosSupportedDistrosSchema().load(byos_supported_distros_schema)
+        assert_that(byos_supported_distros.x86).is_equal_to(x86 or SUPPORTED_OSES)
+        assert_that(byos_supported_distros.arm64).is_equal_to(arm64 or SUPPORTED_OSES)
+
+
+@pytest.mark.parametrize(
+    "template, failure_message",
+    [
+        ("https://template.yaml", None),
+        (None, "Missing data for required field."),
+    ],
+)
+def test_cloudformation_cluster_infrastructure_schema(mocker, template, failure_message):
+    byos_cloudformation_cluster_infrastructure_schema = {}
+    mocker.patch("pcluster.utils.get_region", return_value="fake_region")
+    mocker.patch("pcluster.utils.replace_url_parameters", return_value="fake_url")
+    if template:
+        byos_cloudformation_cluster_infrastructure_schema["Template"] = template
+
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosCloudFormationClusterInfrastructureSchema().load(byos_cloudformation_cluster_infrastructure_schema)
+    else:
+        byos_cloudformation_cluster_infrastructure = ByosCloudFormationClusterInfrastructureSchema().load(
+            byos_cloudformation_cluster_infrastructure_schema
+        )
+        assert_that(byos_cloudformation_cluster_infrastructure.template).is_equal_to(template)
+
+
+@pytest.mark.parametrize(
+    "source, failure_message",
+    [
+        ("https://artifacts.gz", None),
+        (None, "Missing data for required field."),
+    ],
+)
+def test_byos_cluster_shared_artifact_schema(mocker, source, failure_message):
+    byos_cluster_shared_artifact_schema = {}
+    mocker.patch("pcluster.utils.get_region", return_value="fake_region")
+    mocker.patch("pcluster.utils.replace_url_parameters", return_value="fake_url")
+    if source:
+        byos_cluster_shared_artifact_schema["Source"] = source
+
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosClusterSharedArtifactSchema().load(byos_cluster_shared_artifact_schema)
+    else:
+        byos_cluster_shared_artifact = ByosClusterSharedArtifactSchema().load(byos_cluster_shared_artifact_schema)
+        assert_that(byos_cluster_shared_artifact.source).is_equal_to(source)
+
+
+@pytest.mark.parametrize(
+    "artifacts, failure_message",
+    [
+        (["https://artifacts.gz"], None),
+        (["https://artifacts1.gz", "https://artifacts2.gz"], None),
+        (None, "Missing data for required field."),
+    ],
+)
+def test_byos_plugin_resources_schema(mocker, artifacts, failure_message):
+    byos_plugin_resources_schema = {}
+    mocker.patch("pcluster.utils.get_region", return_value="fake_region")
+    mocker.patch("pcluster.utils.replace_url_parameters", return_value="fake_url")
+    if artifacts:
+        byos_plugin_resources_schema["ClusterSharedArtifacts"] = [{"Source": item} for item in artifacts]
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosPluginResourcesSchema().load(byos_plugin_resources_schema)
+    else:
+        byos_plugin_resources = ByosPluginResourcesSchema().load(byos_plugin_resources_schema)
+        for artifact, source in zip(byos_plugin_resources.cluster_shared_artifacts, artifacts):
+            assert_that(artifact.source).is_equal_to(source)
+
+
+@pytest.mark.parametrize(
+    "name, enable_imds, failure_message",
+    [
+        ("user1", True, None),
+        ("user1", None, None),
+        (None, True, "Missing data for required field."),
+    ],
+)
+def test_byos_user_schema(name, enable_imds, failure_message):
+    byos_user_schema = {}
+    if name:
+        byos_user_schema["Name"] = name
+    if enable_imds:
+        byos_user_schema["EnableImds"] = enable_imds
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosPluginResourcesSchema().load(byos_user_schema)
+    else:
+        byos_user = ByosUserSchema().load(byos_user_schema)
+        assert_that(byos_user.name).is_equal_to(name)
+        if enable_imds:
+            assert_that(byos_user.enable_imds).is_equal_to(enable_imds)
+        else:
+            assert_that(byos_user.enable_imds).is_equal_to(False)
+
+
+@pytest.mark.parametrize(
+    "file_path, timestamp_format, failure_message",
+    [
+        ("/var/log/slurmctld.log", None, None),
+        ("/var/log/slurmctld.log", "%Y-%m-%d %H:%M:%S,%f", None),
+        (
+            None,
+            "%Y-%m-%d %H:%M:%S,%f",
+            "Missing data for required field.",
+        ),
+    ],
+)
+def test_byos_file_schema(file_path, timestamp_format, failure_message):
+    byos_file_schema = {}
+    if file_path:
+        byos_file_schema["FilePath"] = file_path
+    if timestamp_format:
+        byos_file_schema["TimestampFormat"] = timestamp_format
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosFileSchema().load(byos_file_schema)
+    else:
+        byos_file = ByosFileSchema().load(byos_file_schema)
+        assert_that(byos_file.file_path).is_equal_to(file_path)
+        if timestamp_format:
+            assert_that(byos_file.timestamp_format).is_equal_to(timestamp_format)
+        else:
+            assert_that(byos_file.timestamp_format).is_equal_to("%Y-%m-%dT%H:%M:%S%z")
+
+
+@pytest.mark.parametrize(
+    "files, failure_message",
+    [
+        ([{"FilePath": "/var/log/slurmctld.log", "TimestampFormat": "%Y-%m-%d %H:%M:%S,%f"}], None),
+        (
+            [
+                {"FilePath": "/var/log/slurmctld.log", "TimestampFormat": "%Y-%m-%d %H:%M:%S,%f"},
+                {"FilePath": "/var/log/scaling.log", "TimestampFormat": ""},
+            ],
+            None,
+        ),
+        (None, "Missing data for required field."),
+    ],
+)
+def test_byos_logs_schema(files, failure_message):
+    byos_logs_schema = {}
+    if files:
+        byos_logs_schema["Files"] = files
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosLogsSchema().load(byos_logs_schema)
+    else:
+        byos_logs = ByosLogsSchema().load(byos_logs_schema)
+        for file, expected_file in zip(byos_logs.files, files):
+            assert_that(file.file_path).is_equal_to(expected_file["FilePath"])
+            assert_that(file.timestamp_format).is_equal_to(expected_file["TimestampFormat"])
+
+
+@pytest.mark.parametrize(
+    "byos_version, events, failure_message",
+    [
+        ("1.0", {"HeadInit": {"ExecuteCommand": {"Command": "env"}}}, None),
+        (None, {"HeadInit": {"ExecuteCommand": {"Command": "env"}}}, "Missing data for required field."),
+        ("1.0", None, "Missing data for required field."),
+    ],
+)
+def test_byos_scheduler_definition_schema(byos_version, events, failure_message):
+    byos_scheduler_definition_schema = {}
+    if byos_version:
+        byos_scheduler_definition_schema["ByosVersion"] = byos_version
+    if events:
+        byos_scheduler_definition_schema["Events"] = events
+    if failure_message:
+        with pytest.raises(ValidationError, match=failure_message):
+            ByosSchedulerDefinitionSchema().load(byos_scheduler_definition_schema)
+    else:
+        byos_scheduler_definition = ByosSchedulerDefinitionSchema().load(byos_scheduler_definition_schema)
+        assert_that(byos_scheduler_definition.byos_version).is_equal_to(byos_version)
+        assert_that(byos_scheduler_definition.events.head_init.execute_command.command).is_equal_to(
+            events["HeadInit"]["ExecuteCommand"]["Command"]
         )